Print a warning when the CLI is out of date (#7)
* Print a warning when the CLI is out of date

This changes adds an update to date check to our CLI. It works by
fetching the latest version of the CLI from a well-known location and
prints a warning if it is larger. To improve user precived
performance, we run the logic to fetch the latest version in parallel
with the command the user is running and we cache the value we fetch
for 24 hours.

If the CLI is out of date, a warning like the following is printed
when the command completes:

```
warning: your version of azd is out of date, you have 0.0.1-beta.1686705 but the latest version is 0.0.1-beta.1686706

To update to the latest version, run:
powershell -c "Set-ExecutionPolicy Bypass Process -Force; irm 'https://aka.ms/install-azd.ps1' | iex"
```

When run on non windows platforms, the `bash` invocation is printed
instead.

This behavior may be disabled by setting AZD_SKIP_UPDATE_CHECK to
`true`.

* Update version URL; set User-Agent

Move to a new format for the aka.ms link for the latest version that
will give us more flexibility when we want to also include channel
information as part of the update check.

Also, set a custom User Agent as we do on other requests that we make.