Add a note to README.md for supported regions
Some of our templates use services which are not supported in all
Azure regions. Long term we hope to encode this information in a way
such that the CLI can restrict the list of locations to offer user to
pick from when they are doing a deployment, but that work is still in
progress. For now, add a note to some of our templates calling out
supported regions.

This list was generated by looking at every RP used for each template
and then fetching information from ARM to see what locations the
resources were suported in. When then intersect all these sets of
locations.

Unsurprisingly, the lists are the same between the node and python
versions of our templates.