Fix debug message when `--help` is passed to CLI
We have logic very early in the startup process of `azd` to do an
initial parse of the command line arguments to see if `--debug` has
been passed (so we can set up our debugging channels to output to the
correct location).

That logic uses `pflag` to parse the command line with a very limited
set of flags (just `--debug`) with it configured to ignore all other
flags. However, pflag treats `--help` as special and when the flag is
not defined returns `ErrHelp` from `Parse` to denote that help was
requested.

Our logic would write any failure from `Parse` to the debug log, which
meant that we would end up printing something like this to stderr at
the start of the command when `--help` was passed:

```
Usage of :
      --debug
```

This change adds a `help` parameter so that we don't hit this behavior
if `pflag`.  The `--help` parameter will be interpreted correctly
later when we parse the command line "for real" using the normal cobra
logic.