Generate combined markdown docs file per build
Our current story for generating docs suitable for ingestion in
docs.microsoft.com is "generate a single markdown file with all the
commands and use anchors to cross reference". Since that's not
something the Cobra markdown formatters support out of the box, PC has
manually been merging these files together and doing some small
tweaks.

This PR attempts to automate that logic.

My first attempt at these changes had us to continue to use the Cobra
logic directly (by importing the package it was in) and then try to
clean up small things with some post processing. However, once I
realized that we would need to re-order the default output of the list
commands from SEE ALSO I decided to just take the existing code and
bring it in here and make small tweaks.

At some point, we will move to a different strategy where we have
multiple files and TOC that stitches them together, but until that
point this frees up a bunch of PC's time.