Update help based on content writer feeback
This commit pushes most of the feedback from the content writer into
the `azd` codebase. There are a few changes we couldn't adopt.

- When listing output formats, we use "json" instead of "JSON" as the
  content writer suggested. This matches what is expected on the
  command line.

- When showing help for a command option with a default value, we
  print `(default: <value>)` at the end of the help text for that
  option. This behavior comes from the library we use for our CLI and
  is not easy to change.

There are a handful of remaning diffs, most of which are due to the
same content being repeated twice and sourced from the same location
(e.g. `provision` and `infra create` share the help text) and the
content writer putting slightly different changes in each place. This
was also common for global options. I've applied some level of
judgement here, but in cases where there was a majority of support for
one form, I adopted that.

There were some edits to the generated markdown that would be hard to
bring back into the code (espically around URL formatting, where the
markdown reads very poorly from a CLI). In these cases I did my best
to make it look nice on the command line first and then to have a
reasonable rendering in the generated markdown second.