Improve `azd` documentation story (#116)
* Generate combined markdown docs file per build

Our current story for generating docs suitable for ingestion in
docs.microsoft.com is "generate a single markdown file with all the
commands and use anchors to cross reference". Since that's not
something the Cobra markdown formatters support out of the box, PC has
manually been merging these files together and doing some small
tweaks.

This PR attempts to automate that logic.

My first attempt at these changes had us to continue to use the Cobra
logic directly (by importing the package it was in) and then try to
clean up small things with some post processing. However, once I
realized that we would need to re-order the default output of the list
commands from SEE ALSO I decided to just take the existing code and
bring it in here and make small tweaks.

At some point, we will move to a different strategy where we have
multiple files and TOC that stitches them together, but until that
point this frees up a bunch of PC's time.

* Update help based on content writer feedback

This commit pushes most of the feedback from the content writer into
the `azd` codebase. There are a few changes we couldn't adopt.

- When listing output formats, we use "json" instead of "JSON" as the
  content writer suggested. This matches what is expected on the
  command line.

- When showing help for a command option with a default value, we
  print `(default: <value>)` at the end of the help text for that
  option. This behavior comes from the library we use for our CLI and
  is not easy to change.

There are a handful of remaining diffs, most of which are due to the
same content being repeated twice and sourced from the same location
(e.g. `provision` and `infra create` share the help text) and the
content writer putting slightly different changes in each place. This
was also common for global options. I've applied some level of
judgement here, but in cases where there was a majority of support for
one form, I adopted that.

There were some edits to the generated markdown that would be hard to
bring back into the code (especially around URL formatting, where the
markdown reads very poorly from a CLI). In these cases I did my best
to make it look nice on the command line first and then to have a
reasonable rendering in the generated markdown second.