Infra Provision/Deploy provider model (#172)
The is the start of the new infrastructure provisioning provider model. This is primarily net new code with a couple minor updates to existing files.

The goal of this PR is to abstract much of the infrastructure provisioning out into a provider interface with a new Bicep implementation. This will need to be tweaked as we continue to understand the differences between different infra providers like terraform, bicep, etc.

Contains the following:

 New Bicep infra provider implementation
 New async task model to make it easier to consume long running tasks/operations with progress reporting
 New approach to mocking and setting up mock definitions that call into dependent executil compatible tools
 New infrastructure provisioning manager that will greatly reduce the amount of code required in the infra create/delete commands