Reuse spinner instance when available (#591)
With the introduction of provisioning manager it is now possible that there could be multiple instances of spinner created within different components. This PR introduces the spinner into the context where it can be leveraged by downstream components.

Without reusing the spinner component we see flashing display as the 2 spinner instances attempt to overwrite each other.

Ex) Spinner is used in the deploy command that calls into the provisioning manager.