Use `type` not `language` in show output
This makes the set of values smaller and doesn't use the word
"language" which can be sort of confusing when it's set to something
like "dotnet".

As part of this, we normalize the output so that `type` is always one
of `dotnet`, `python` or `node` to make consumers lives easier.