Move project sniffing logic into `show`
Consumers like Visual Studio need to understand the project file that
is assocated with a dotnet service, not just the path to the directory
the service lives in.

Initially, this logic was in it the project reader (which builds a
ProjectConfig), but that felt like the wrong location for the
logic. We can't add this logic to `ProjectConfig::GetProject` because
that call fails when the infrastructure for an appliaction has not
been provisioned, and we want `azd show` to work even in that case.

Longer term we should look into something that allows `GetProject` to
work evern when infrastructure is not created, and at that point move
this logic there, but for now we have this as a specific
implementation detail of the show command.