Embrace `*Environment` over `Environment` (#648)
Today the `Environment` object just holds a go map of values which
represent the environment. Different parts of the system had different
opinions as to if a pointer to an environment or the entire object
should be passed around.  We intend to chage the internal shape of
this object, so to prepare, I've done a pass over the codebase to do
the following:

- Environments are now only created by code in the environment package,
  and all the "constructors" return `*Environment`.

- `Ephemeral` and `EphemeralWithValues` have been introduced to the
   environment package for testing purposes. They can be used to
   replace explicit construction of environments.

- Methods that return an environment along with an error now return
  `nil`, instead of the zero value for `Environment` (which is itself
  invalid since the map of values is nil) to prevent accidently use.

As part of this change, I noticed a few places where we would load an
environment and then not pass it to our project reading logic (in
favor of using an invalid one with no values). For `restore` this
required actually loading the environment, which we didn't do
previously.