Support more `docker --version` schemes
Docker EE has been replaced with Docker CE / Moby for the windows
images used by both GitHub Actions and the ones we use for our
internal AzDo runs.

With this change, docker --version started to report a version number like:

```
Docker version master-dockerproject-2022-03-26, build dd7397342a
```

Our logic to parse the version number could not handle this scheme. It
has now been rewritten to support this, as well as the older scheme
which is still used by the Mac and Linux images (as well as for folks
who have installed Docker Desktop on Windows)