Contractify `env refresh`'s JSON output
When we added support for terraform, we changed the JSON output from
`azd env refresh` somewhat unintentionally. Previously, it had been
the full deployment object, from ARM, but since there is no deployment
object when using Terraform, we could not long return that complete
value. What we ended up returning was what the provider interface
calls a `Deployment` which is a set of parameters and outputs, with no
resources.

This change moves the output of `env refresh` to be a type we have a
specific contract for and right now contains just the outputs and
resources from the most recent infrastructure update. For ARM based
deployments, this information is sources from the most recent
deployment object and for terraform it is discovered by reading the
current state file (via `terraform show`)

Contributes To: #742