Adopt `t.Setenv` in more of our tests
Since `azd` can be controlled via environment variables, many of our
tests end up setting environment variables. Tests had different
strategies on how this was done and how things were cleaned up
afterwards, but the standard testing interfaces provide a nice
primative `t.Setenv` which ensures that the test does not run in
parallel with other tests (important, since this is global state we
are modifying) and ensures the value is restored after the test has run.

This change adopts it across our tests. For cases where we need to do
something more complex, we have some helpers in `ostest` that can be
useful, including a missing `t.Unsetenv` like thing we'd want to use a
few places.

To prevent an import cycle, I moved the code in `ostest` that deals
with cleaning up a directory and sending telemetry to us in CI when it
fails into the `cli_test.go` file, which is the only place it was
used.