Rework provider implementation strategy
Move to using an `init` function on each of our packages that provides
an implementation of the `provisioning.Provider` interface. This
allows us to do this registration before any user code is running
which allows us to drop that was protecting the map of providers from
concurrent modification.

Places which used to call `RegisterProvider` can now use a discard
import to acomplish the same thing.