Move away from `fmt.Print` and friends (#879)
Instead of using `fmt.Print`, which forces output to `os.Stdout`, this
change moves us so that we use the `fmt.Fprint` family of functions,
to allows us to control where output goes.

In places where we already have a `cobra.Command`, I have used
`cmd.OutOrStdout()` out and friends to get the output stream we want
to write to. For other components, the `Console` type now has a
`Handles()` property that you can use to access Stdout, Stderr and
Stdin.

This change also removes the ambient use of `os.Stdout`, `os.Stdin`
and `os.Stderr` in as many places as possible. The only remaining
locations are for setting up our context objects and some code in our
process runner for the case where we are running an "interactive"
command and we want to have our standard streams be used by the
child. That will be cleaned up in a future change.