SetWriter should restore to original writer
I had intended this to be part of #891 but forgot to commit before
pushing the PR. Originally, when SetWriter(nil) was called, we would
use `GetDefaultWriter()` (which has since been removed) to re-fetch
the same value that was passed in to `writer` when the object was
created.

So store the originial writer for now and restore to it. We could
otherwise run into issues when NO_COLOR was set where we'd not restore
the stream that striped the ANSI color codes.