Improve `--output="json"` behavior of `azd`
To support their extension, Visual Studio calls `azd` with
`--output=json` and uses these structured messages to drive behavior
internally.

Previously, when running `azd infra create --output=json` we would
spit out an array of deployment operations peridocially and then at
the end of the operation we would print he full deployment object from
ARM.

This behavior broke in subtle ways when did the refactoring to support
multiple IaC providers, and that turned up an ever larger issue - the
entire contract was very tied to an ARM deployment. When using
Terraform there is no set of ARM Deployment Operations we can provide
(since it doesn't use ARM Deployments, instead it just calls the
control plane directly) and there's no deployment object at the end
that we can return.

This change works to improve this behavior.

Our contract types now include a top level `type` property that
consumers can use to understand the shape of the message. This was
added to allow us to emit more messages as the CLI evolves for a given
operation. Consumers should ignore objects with types the don't
understand.

Calls to `console.Message` are now written in a structured format
when using `--output=json`. Any ANSI color codes are removed from the
output.  New lines are kept as is, so a consumer can take any object
with the `type` of `consoleMessage` and simply JSON decode the
`message` property and then write it to the console (any required
trailling newlines will be part of the message string) which will give
them the same output that would have been rendered the the console.

When consuming output from the CLI run with `--output=json` consumers
should read JSON objects from stdin until EOF is reached.  Objects
will be streamed in "real time" as the CLI produces them.

For example, here's the output of `azd infra create --output=json`
with these changes for a small App Service application:

```
{
  "type": "consoleMessage",
  "message": "\nInfrastructure provisioning plan completed successfully\n"
}
{
  "type": "consoleMessage",
  "message": "Provisioning Azure resources can take some time.\n\nYou can view detailed progress in the Azure Portal:\nhttps://portal.azure.com/#blade/HubsExtension/DeploymentDetailsBlade/overview/id/%2Fsubscriptions%2Ffaa080af-c1d8-40ad-9cce-e1a450ca5b57%2Fproviders%2FMicrosoft.Resources%2Fdeployments%2Fmatell-azd-json\n\n"
}
{
  "type": "consoleMessage",
  "message": "Created Resource group: rg-matell-azd-json\n"
}
{
  "type": "consoleMessage",
  "message": "Created App Service plan: plan-itzzednmw27ng\n"
}
{
  "type": "consoleMessage",
  "message": "Created Web App: app-itzzednmw27ng\n"
}
{
  "type": "consoleMessage",
  "message": "\nAzure resource provisioning completed successfully\n"
}
{
  "type": "consoleMessage",
  "message": "View the resources created under the resource group rg-matell-azd-json in Azure Portal:\nhttps://portal.azure.com/#@/resource/subscriptions/faa080af-c1d8-40ad-9cce-e1a450ca5b57/resourceGroups/rg-matell-azd-json/overview\n\n"
}
{
  "type": "envRefreshResult",
  "outputs": {
    "WEBSITE_URL": {
      "type": "string",
      "value": "https://app-itzzednmw27ng.azurewebsites.net/"
    }
  },
  "resources": [
    {
      "id": "/subscriptions/faa080af-c1d8-40ad-9cce-e1a450ca5b57/resourceGroups/rg-matell-azd-json"
    },
    {
      "id": "/subscriptions/faa080af-c1d8-40ad-9cce-e1a450ca5b57/resourceGroups/rg-matell-azd-json/providers/Microsoft.Web/serverfarms/plan-itzzednmw27ng"
    },
    {
      "id": "/subscriptions/faa080af-c1d8-40ad-9cce-e1a450ca5b57/resourceGroups/rg-matell-azd-json/providers/Microsoft.Web/sites/app-itzzednmw27ng"
    },
    {
      "id": "/subscriptions/faa080af-c1d8-40ad-9cce-e1a450ca5b57/resourceGroups/rg-matell-azd-json/providers/Microsoft.Web/sites/app-itzzednmw27ng/config/appsettings"
    }
  ]
}
```

Contributes To: #742