Rework progress streaming with `--output=json`
Based on discussions with Wei and Phillip, we've decided to make some
changes to how we report progress when `--output=json` is set.

In the previous commit, we would write progress information as well as
the final result of the command to stdout. This meant that we needed
to have a `type` parameter on all the JSON data we emittied so that
consumers could understand how to decode things. Furthermore, we
decided to wrap these progress messages in an envelope so that the
metadata and data were seperated and we wouldn't run into collisions
where we a property from the data we wanted to emit clashed with a
property name for some metadata we needed to provide to
consumers. This strategy was workable, but it lead to sort of bad UX
on the command line, as the result of every command now looked like:

```jsonc
{
  "type": "someEventType",
  "data": // a JSON value that represents the actual data
}
```

Which meant consumers would always need to access the `data` property
to get the information they were interested in.

Furthermore, this cost was going to be paid by all commands (not just
ones that emit progress) so that our output was regular and also to
give us the flexability to return more status information over time.

After discussion, we decided to instead adopt the following model:

When a command is run with `--output=json` the result of the command,
in JSON form is printed to stdout. Since only a single value will ever
be permitted, there is no need to have an envelope since the consumer
knows, a priori the shape of the result given the command they are
executing.

During an operation, when `--output=json` is specified, the CLI may
also emit a stream of events to stderr which contain information about
the operation that is in progress.  These events are wrapped in an
envelope (which includes both a type property, telling consumers the
shape of the data and a timestamp as to when the event was
emitted. Consumers may ignore the stderr stream completely or they may
choose to read JSON objects, one at a time, from the stream and decode
them if they understand the `type` of the event.  Over time, the CLI
may start to emit more events over this output stream, so consumers
SHOULD ignore events with `type`'s they don't understand.

We also discussed how we will want to report errors from commands when
`--output=json` is true. We decided that in the case of a fatal error,
we'll write a structured error object to stdout (instead of the normal
result of the command, which we don't have). Consumers will need to
look at the exit code of `azd`. When zero, they can interpret the
object written to stdout in the usual way, but when non zero, they
should instead assume the object represents an error and they can
decode it in that manner. Wei is working on the shape of the error in
that case.