Write JSON objects to `stderr` without indenting.
Wei and I chatted about this on a Teams call. His concerns with our
existing strategy of calling Format (and hence getting indented JSON
objects spanned over multiple lines pushed to stderr) were twofold:

1. Looking across the ecosystem, it seemed like most tools would write
these diagnostics messages as a single line and it seemed weird for us
to divert.

2. In cases where you may have a JSON reader that works on strings,
not streams, it may be harder to read set of objects from a string,
since these libraries might only expect a single object (since they
take a string). If each object was on its own line, you could split
the string on newlines to get an array of strings, each which contains
a single object, and that would be easier to parse with these
libraries.

Given that these events are targeted at machines, not humans, we
decided it would be OK to put them on a single line instead of
formatting them nicely.

We'll continue to format the final result with newlines and
indentation, since that's much nicer UX for folks and also we don't
run into the above problem since there would only ever be a single
object written to stdout.