Simplify color processing on Windows
Historically, the Windows console has used out of band signaling for
control color information, unlike *nix where control characters are
added to the text to control what color it is printed in and other
formatting information.

For a long time, Windows could not cope with the *nix strategy,
leading to things like `colorable.NewColorableStdout()` and friends
which tried to provide streams which would interpret ANSI escape
sequences and call the corresponding Win32 methods to change colors as
it rendered the text.

This process was very fragile and would lead to cases where we'd end
up not using `colorable.NewColorableStdout()` and then when run on
windows in the legacy console host we'd see these escape codes get
printed.

Windows now supports processing these escape sequences. Enable that
behavior and get out of the `colorable.NewColorableStdout()` game.

Fixes #1058