Mark more commands as not requiring login
As fallout of #1022 we ended up requiring authentication for some
commands that did not need it. This change ads annotations to these
commands such that they do not require authentication.

We should be able to remove these annotations and have everything be
correct by construction when we stop putting a credential and azcli on
the context but instead just inject these into the commands that need
them (and only the commands that need them).

Fixes #1135