Mark `pipeline config` as requiring login
This command requries that you are logged into Azure to work, but in

I removed the incorrect annotation, and I also added a `azcli.AzCli`
parameter to the type that represents the action itself and used wire
to ensure it is injected. This is work we'd do any way as part of
remvoing `commands.RequireNoLoginAnnotation` (see #1145) so might as
well lock it is now.

Fixes #1147