Make `-e` flag non-global (#1143)
`azd` currently has `-e` or `--environment` flag marked as a global flag, which is often times just slightly confusing due to it not taking-effect for commands that are not environment-scoped, like `azd login` or `azd config`.  Moreover, for commands that also operate on the environment verb `azd env select` and `azd env new`, this is confusing and harmful.

This change removes `-e` as a global flag and attaches it to commands that operate under the environment concept. 

Addresses #1109