Fix issues so `todo-csharp-cosmos-sql` can be run locally
The `todo-csharp-cosmos-sql` template had two issues that prevented it
from being used locally:

1. It did not set `AZURE_COSMOS_ENDPOINT` as an output of the
infrastructure - but this value was used by the application to at
runtime to connect to the cosmos database. When run in App Service we
correctly defined the environment variaible, but since we didn't
export it from the infrastructure, it was not added to the `.env` file
and hence not avaiaible for the app to use when running locally.

2. It did not grant RBAC access to the local user, so once (1) was
fixed, the local user could not access the database.

This change fixes both of these issues.

Fixes #1101