Fix prompting for missing infrastructure values
Our logic to prompt the user for missing infrastructure values was
broken because we did not tell the spinner we wanted to interact with
the user, so the spinner would prevent the prompt UI from being shown.

The logic we had for trying to save these values was totally broken
and in practice would end up generating files that the ARM control
plane would reject.

We now save these values in the newly added `Config` object on the
environment, using `infra.prameters.<parameterName>` as the key name.

As part of this work, we no longer materialize the full
`.parameters.json` file to disk, we no longer need to do this now that
we are using the SDK directly.

Fixes #1176
Fixes #1191