Remove `RegisterDependenciesInCtx` (#1195)
This change does the majority of the work to stop using `context.Context` as an IoC container, in favor of explicitly injecting dependencies where needed (automated by `wire` at top level).

We still have a few types that have `.Get` and `.With` methods that allow registering and retrieving values from the context:

- `httputil.HttpClient`
- `spin.Spinner`
- `config.Manager`
- `bagage.Baggage`

I plan to remove the first three in a follow up PR.

I think that `baggage.Baggage` should remain for now because it's more tied to our telemetry/tracing system which logically builds upon the context object (@weikanglim, does this seem right?) - but also looking at the code it's not clear to me if we are still using this stuff at all.  It seems like the `ContextWithBaggage` and `ContextWithoutBaggage` are both unused outside of test code.  Perhaps this is related to #1190?

Finally, we have the `tools.WithInstalledCheckCache(context.Context) context.Context` helper which we use to register a cache to speed up our install checks.  This is just now used directly in `root.go`.  I think we may want to revisit this pattern at some point (perhaps after thinking more about how we want our external dependencies and wire to interact, it may be possible that we can refactor this stuff such that we no longer have `CheckInstalled` but rather commands are injected the tools they need, and wire validates they are installed as part of construction) but right now this feels reasonable to me.  In practice we don't export at `GetInstalledCheckCache` method, so arbitrary components can't take a dependency this.