Have `azd` acquire Bicep
This change updates our behavior so that `azd` uses a copy of `bicep`
that it manages itself, similar to how `az` worked.

When using Bicep, we now check for a copy of bicep in
`$AZD_CONFIG_DIR/bin` and if it is not there (or it is there, but is
older than our minimum supported version) we download a copy.

As a safety valve, one may set `AZD_BICEP_TOOL_PATH`. If that's done,
this value is used as the path to Bicep and no version validation is
done.