Fix regression on loading existing environments
In #1202 we added a `config.json` file in the environment, which we
use to persist configured values for required infrastructure
parameters which are not set in `.parameters.json` files.

As part of that logic, we made it so that the constructor of the
environment would validate that a `config.json` file was present in
the environment and if not, fail (similar to the check it already did
for the .env file).

This worked for new environments, since `.Save()` will always write an
empty `config.json` file. But for existing environments, the read of
the config.json would fail, with ErrNotExists, which would be wrapped
up and returned to the caller. The caller would use the ErrNotExists
to mean "the environment doesn't exist" and so it would prompt the
user to see if they wanted to create it.

Rework the constructor such that we return ErrNotExists when the
directory for the environment doesn't exist, instead of using the
existence of the `.env` file to mean anything and add some unit tests
for these cases.