Improve prompting for locations
With this change, we now support adding the following metadata to a
`string` parameter:

```
{ azd: { type: 'location' } }
```

For example:

```bicep
@description('Primary location for all resources')
@metadata({ azd: { type: 'location' } })
param location string
```

When `azd` prompts for a string value with type set to "location" it
will present a location selection box instead of asking the user for a
raw string value. The value of the parameter will be assigned to the
short name of the location (e.g. "eastus2").

This composes with the `@allowed` decorator - when @allowed also also
present, the full set of locations the user can reach is filtered
against this list (it should contain the short names of the regions
that should be allowed).