Delay resolving `${var}` references when loading azure.yaml (#1276)
We support using `${var}` style substitutions in `azure.yaml` today,
but the current implementation is very simple: we do a single
substitution pass before attempting to parse the YAML that makes up
the file.

This means to get a `ProjectConfig` instance, you need an
`*Environment` and that's something we can't fix with the given design
because the yaml file may not even be valid until we do
substitutions (and hence we couldn't parse it to form the ProjectConfig).

After chatting about this with Wei, We'd like the remove the
dependency of ProjectConfig on `*Environment` and at the same time be
more restrictive around where these references could appear (for
example, in the case of the inline script work Wallace is prototyping,
the current behavior would cause references in lines scripts that look
like `${var}` to be replaced while the file is being loaded, which is
not correct).

This change pushes this idea into the codebase.  Looking at the
existing fields of `ProjectConfig` it seemed like we should really
only allow substitutions for the name of the primary resource group
and in the tag parameter for docker containers (where we documented
that we support substitutions). We can always add more stuff later but
this seemed like a reasonable place to start.