Rework parameters in Bicep Provider (#1222)
This set of changes improves our prompting behavior for required
deployment parameters for bicep based deployments.

The following improvements have been made:

- We now correctly support all the bicep types (strings, arrays,
  objects, ints, bools).

- The Description metadata of a parameter is used at the help text
  during the prompt, if it is provided in the template.

- When prompting for a parameter of type "bool", we provide a
  selection list which contains "true"/"false" instead of asking the
  user to type in a string.

- When prompting for a parameter which has the `allowedValues`
  metadata set, a list is presented of the allowed values and the user
  can choose one instead of asking the user to type in a string.

- We now support all the validators (`minLength`, `maxLength`,
  `minValue`, `maxValue`) for string and int types.

- ARM templates can now contain an `azd` object in the `metadata`
  section, which provides azd specific metadata that can be used to
  improve prompting. This is a map[string]string today and we support
  one setting.  The `type` key can be set to `location` and if done it
  signifies that the value for the parameter should be the short name
  of an azure region.  When prompting, we use our location selection
  flow instead of asking a user to put in something like `eastus2`.
  This composes with `allowedValues` allowing a template author to
  restrict the set of allowable locations (useful in cases where a
  service is not supported in different regions).

In addition, some changes have been made to how the bicep provider
stores these values between plan and apply. Previously they were
stored as values on the `Deployment` object's parameters array, which
also stored information about all the parameters from the template we
were going to deploy. Now, instead, they are stored inside the bicep
specific DeploymentDetails object that is attached to the plan (along
with the JSON of the template we are going to deploy).  This allows us
to stop mixing the definition of parameters with their assignments
which more correctly follows how bicep models things.