Don't print an extra newline after (most) commands (#1297)
If a command returned a nil ActionResult (as most commands do), we
would still try to create a MessageUxItem for it and render that. This
would end up with an empty message being printed to the console and
then another newline. When using JSON output this was very obvious
(when running at the terminal, it's a little easy to not notice the
extra blank line that is at the end of the command).