Allow configuration of service endpoint using `SERVICE_<service>_ENDPOINTS` (#1312)
Allow users to specify their own endpoints for a given service, in cases where they've configured their own front-end load balancers, reverse proxies or DNS host names outside of the service target (and prefer that to be used instead). Uptake this change for the APIM-enabled template.

As part of this change, `azd` now serializes complex type outputs from the infrastructure provider as JSON strings when storing in the environment. Previously, the output was stored as `golang`-specific string type representations.

Fixes #1192