Pass the correct context from middleware to the action (#1330)
In the case where we had reached the end of the middleware chain, we
used an outdated context. If the middleware chain changed the context
object (as the telemetry middleware does, when it creates a new span)
we'd lose these updates. In the case of telemetry, it means that any
child spans we created would not be correctly parented.