Support writing Open Telemetry compatible traces (#1327)
This change adds two new (hidden) flags to `azd`: `--trace-log-file`
and `--trace-log-url`. These flags can be used to export tracing
information in the open telemetry format either to a file or to an
HTTP collector.

Right now these flags are hidden because they are only of use to us as
developers for `azd` but longer term we might expose
`--trace-log-file` as a way to have users capture traces while
reporting bugs. Right now it's most useful for developers who want to
improve our telemetry since exporting to a local HTTP collector and
inspector (like Jaeger) provides a tight feedback loop.