Fix `azd up --template` so it can initialize a new project
In #1296 we changed our logic such that we would inject `AzdContext`
as a depenecy of our actions, and expect our IoC container to wire
things up.

This had the side effect of breaking `azd up --template` to initilize
(and then provison and deploy) a new project.

The break comes from the fact that the IoC container will call
`NewAzdContext` as part of building the `deploy` and `infraCreate`
actions, which need to be created because they are dependencies of the
`up` composite action.  However, `NewAzdContext` should not be called
before the project has actually been created (which will happen when
the `up` composite action calls the `init` action), because it looks
for an existing project and if it doesn't find one it fails.

To work around this issue - I've made the infra create and deploy
actions explicitly call `NewAzdContext` so the calls can happen at the
right time.

A regression test has been added (it's a little hacky because we don't
actually care about running the `infra create` or `deploy` parts of
`up` in this test, we just want to ensure that we correctly
initialized via a template.

Fixes #1329