Avoid loading entire user profile for `homeDir` (#1470)
This switches `user.Current().HomeDir` to `os.UserHomeDir`, avoids loading entire user profile when only the home directory is needed, and improves `azd` (including `--help` only) performance by ~100ms to seconds on certain Windows machines.

`user.Current` fetches the entire user profile (this is loaded once and then cached). On *Nix, it simply reads pwd through a single syscall. However, on Windows, it requires talking to the domain controller for a domain-joined account. This is on magnitude of ~100ms to seconds slow which is noticeable, and also impacts the `--help` experience (since we currently have logic that runs in `main` irrespective of flags).

This changes the underlying implementation to use %HOMEPROFILE% or $HOME env variables instead of looking the system user. This should not be a problem and now matches how `azd config` is documented, where %HOMEPROFILE% or $HOME is stated explicitly.

Thanks to @bwateratmsft for reporting this.