Correctly show function app resources in provision log
Function Apps and App Service apps use the same Azure resource types,
so our display logic needs to query information about the provisioned
resource to understand if a provisioned resource is an app service or
a function app.

Our existing logic to do this was broken when we moved away from using
`az` to get information about a resource, since we were not passing an
API version on our query. This caused our REST call to fail, and we
just ignored the error (under the assumption it was a transient issue
and we were okay displaying slightly wrong output for progress in this
case).

This change addresses the issue by forcing callers (who in theory
understand what API version is valid for the resource they want to
fetch information about) to provide an API verison and updates the one
call site we have to pass a valid API version.

Without this change, when deploying a template like
`todo-nodejs-mongo-swa-func` you'd see this in the provision log:

```
(✓) Done: Web App: func-api-trxnle2vswr24
```

Whereas with this change, we now correctly report that this is a
function app:

```
(✓) Done: Function App: func-api-trxnle2vswr24
```

Fixes #1484